% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaverfunctions.r
\name{activate_colorout}
\alias{activate_colorout}
\title{activate_colorout}
\usage{
activate_colorout()
}
\description{
Syntax Highlighting
}
\details{
Install and activate terminal syntax Highlighting
}
\author{
Ben Weaver
}

