% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaverfunctions.r
\name{activate_packages}
\alias{activate_packages}
\title{activate_packages}
\usage{
activate_packages(...)
}
\arguments{
\item{pkg_vect}{A vector of character strings}
}
\value{
logical
}
\description{
Install and load multiple packages
}
\details{
Takes a vector of character package names and loads or installs then loads the named packages
}
\author{
Ben Weaver
}

