% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WeaverFunctions.r
\name{geom_mean}
\alias{geom_mean}
\title{geom_mean}
\usage{
geom_mean(x)
}
\arguments{
\item{x}{A vector of numerics or integers}
}
\value{
numeric
}
\description{
Geometric Mean
}
\details{
Performs geometric (product) mean on a vector
}
\author{
Ben Weaver
}

