% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaverfunctions.r
\name{get_directory_file_names}
\alias{get_directory_file_names}
\title{get_directory_file_names}
\usage{
get_directory_file_names(directory)
}
\arguments{
\item{directory}{A character vector of file name paths}
}
\value{
character
}
\description{
Get containted file paths from directory path
}
\details{
Takes directory path and returns contained file paths
}
\author{
Ben Weaver
}

