% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaverfunctions.r
\name{gzfread}
\alias{gzfread}
\title{gzfread}
\usage{
gzfread(path, sep = "\\t", out_dir = NULL)
}
\arguments{
\item{path}{a character}
}
\value{
data.table
}
\description{
Import text from gzipped or gunzipped file to data.table
}
\author{
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}

