% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaverfunctions.r
\name{install_missing_package}
\alias{install_missing_package}
\title{install_missing_package}
\usage{
install_missing_package(pkg)
}
\arguments{
\item{pkg}{A character}
}
\value{
logical
}
\description{
Install or load a package
}
\details{
Takes a package name as a character string and loads or installs then loads the named package
}
\author{
Ben Weaver
}

