% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaverfunctions.r
\name{merge_files}
\alias{merge_files}
\title{merge_files}
\usage{
merge_files(file_paths)
}
\arguments{
\item{file_paths}{a character}
}
\description{
Merge files with same header into subset file
}
\details{
Merge files with same header into subset file
}
\author{
Ben Weaver
}

