% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaverfunctions.r
\name{qwritesamples}
\alias{qwritesamples}
\title{qwritesamples}
\usage{
qwritesamples(dir, nsamples = 10000, seed = NULL)
}
\arguments{
\item{dir}{A character directory path}

\item{nsamples}{An integer number of rows to sample from each file}

\item{seed}{An integer seed for repeatable samples}
}
\description{
Get samples of files in directory
}
\details{
Write N samples of each file in directory (separate sample file for each file)
}
\author{
Ben Weaver
}

