% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaverfunctions.r
\name{split_file}
\alias{split_file}
\title{split_file}
\usage{
split_file(file_path, split_col_name)
}
\arguments{
\item{file_path}{a character}

\item{split_col_name}{a character}
}
\description{
Split file into subset files by column name
}
\details{
Split file into subset files by column name
}
\author{
Ben Weaver
}

