% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WeaverFunctions.r
\name{activate_colorout}
\alias{activate_colorout}
\title{activate_colorout}
\usage{
activate_colorout()
}
\description{
Syntax Highlighting
}
\details{
Install and activate terminal syntax Highlighting
}
\author{
Ben Weaver
}

