% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WeaverFunctions.r
\name{chunk}
\alias{chunk}
\title{chunk}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{A vector}

\item{n}{number of chunks}
}
\value{
list
}
\description{
Chunk vector in N equal sizes
}
\details{
Chunk vector in N equal sizes and return list of sub vectors. Taken from StackOverflow example
}
\author{
Ben Weaver
}

